Create blog header / footer / content styles
This will create development folder/file structure with static files
which can be used for ongoing development (makes the process
quicker/easier) and correctly setup up font-awesome in the project.

Apart from the above minor setup, this will define the layout & design
of the blog, create a home, single-post and blog index page. Include
hooking up the style to Jekyll based layouts and includes so it
functions as a Jekyll based site.

Closes issue #3