Convert to a ready-to-use standalone github pages template
Made amendments to the site across the board to be able to quickly
adapt the template to create other github pages site. Added instructions
to the _config.yml to specify which section need to be modified and
similarly added instructions to the variables.css file.

Edited whole site to have url's prepended with the baser so that links
will work as user/organisation site or even as a project site.

Edited the README.md file with instructions on how to adapt this as a
template for another users/projects GitHub pages site.

Closes issue #10